// src/app/seaexport/models/Tbl_cargo_exp_container.ts
export interface Tbl_cargo_exp_container {
    cntr_pkid: string;
    cntr_hblid: string;
    cntr_catg: string;
    cntr_order: number;
    cntr_no: string;
    cntr_type: string;
    cntr_sealno: string;
    cntr_movement: string;
    cntr_pieces: number;
    cntr_packages_uom: string;
    cntr_packages: number;
    cntr_cbm: number;
    cntr_teu: number;
    cntr_weight_uom: string;
    cntr_weight: number;
    cntr_tare_weight: number;
    rec_year: number;
    cntr_rider: string;
    cntr_yn: string;
    cntr_selected:boolean;
}